# packages/spec/openapi.yaml
openapi: 3.1.0
info:
  title: Bark & Meow — API
  version: 1.0.0
  description: |-
    Contrato de apps/api: lo consumen el portal del dueño (mi-mascota), la app, la web de la clínica, la web del veterinario, el panel del operador y el software de gestión.

    Dos principios que cualquier implementación debe respetar:

    1. **El servidor no descifra.** Todo lo que lleva contenido clínico viaja como bytes opacos en base64.
    2. **Localizar no es abrir.** La consulta por identificador responde si existe ficha y nada más, con el mismo tamaño y el mismo tiempo en ambos casos.

    La consulta entre nodos (`/federation/v1/lookup`) está por implementar: solo podrá decir si existe ficha y ofrecer un enlace de aviso, con peticiones y respuestas firmadas con Ed25519.

    Generado desde packages/schema/src/api. No se edita a mano.
  license:
    name: CC BY 4.0
    url: https://creativecommons.org/licenses/by/4.0/
servers:
  - url: https://barkandmeow.app/api
    description: Producción
  - url: http://127.0.0.1:4601
    description: Desarrollo
paths:
...
  /ops/v1/claims:
    get:
      tags:
        - Operador
      summary: Reclamaciones abiertas e impugnadas y las resueltas en los últimos 30 días. Sin OPS_TOKEN válido responde 404, como una ruta desconocida.
      security: &a5
        - tokenOperador: []
      responses:
        "200":
          description: Correcto
          content:
            application/json:
              schema:
                $schema: https://json-schema.org/draft/2020-12/schema
                type: object
                properties:
                  reclamaciones:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          format: uuid
                          pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                        petId:
                          type: string
                          format: uuid
                          pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                        reclamantePetId:
                          type: string
                          format: uuid
                          pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                        clinicId:
                          type: string
                          format: uuid
                          pattern: ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
                        estado:
                          type: string
                          enum:
                            - abierta
                            - impugnada
                            - a-favor-reclamante
                            - a-favor-titular
                        motivo:
                          type: string
                          enum:
                            - impugnada
                            - vencida-sin-cuenta
                            - en-plazo
                            - resuelta
                        plazo:
                          type: string
                          format: date-time
                          pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d:[0-5]\d(?:\.\d+)?(?:Z|([+-](?:[01]\d|2[0-3]):[0-5]\d)))$
                        creada:
                          type: string
                          format: date-time
                          pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d:[0-5]\d(?:\.\d+)?(?:Z|([+-](?:[01]\d|2[0-3]):[0-5]\d)))$
                        resuelta:
                          anyOf:
                            - type: string
                              format: date-time
                              pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d:[0-5]\d(?:\.\d+)?(?:Z|([+-](?:[01]\d|2[0-3]):[0-5]\d)))$
                            - type: "null"
                        nota:
                          type:
                            - string
                            - "null"
                        clinica:
                          type: object
                          properties:
                            nombre:
                              type: string
                            pais:
                              type: string
                            direccion:
                              type:
                                - string
                                - "null"
                            registroSanitario:
                              type:
                                - string
                                - "null"
                            dominio:
                              type:
                                - string
                                - "null"
                            verificada:
                              anyOf:
                                - type: string
                                  format: date-time
                                  pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d:[0-5]\d(?:\.\d+)?(?:Z|([+-](?:[01]\d|2[0-3]):[0-5]\d)))$
                                - type: "null"
                          required:
                            - nombre
                            - pais
                            - direccion
                            - registroSanitario
                            - dominio
                            - verificada
                          additionalProperties: false
                        titular:
                          type: object
                          properties:
                            estado:
                              type: string
                              enum:
                                - pendiente
                                - activa
                                - congelada
                                - retirada
                            chipPista:
                              type:
                                - string
                                - "null"
                            activada:
                              anyOf:
                                - type: string
                                  format: date-time
                                  pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d:[0-5]\d(?:\.\d+)?(?:Z|([+-](?:[01]\d|2[0-3]):[0-5]\d)))$
                                - type: "null"
                            activadaPor:
                              type:
                                - string
                                - "null"
                            registrada:
                              type: string
                              format: date-time
                              pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d:[0-5]\d(?:\.\d+)?(?:Z|([+-](?:[01]\d|2[0-3]):[0-5]\d)))$
                            nombre:
                              type:
                                - string
                                - "null"
                            correo:
                              type:
                                - string
                                - "null"
                            correoVerificado:
                              anyOf:
                                - type: string
                                  format: date-time
                                  pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d:[0-5]\d(?:\.\d+)?(?:Z|([+-](?:[01]\d|2[0-3]):[0-5]\d)))$
                                - type: "null"
                          required:
                            - estado
                            - chipPista
                            - activada
                            - activadaPor
                            - registrada
                            - nombre
                            - correo
                            - correoVerificado
                          additionalProperties: false
                        reclamante:
                          type: object
                          properties:
                            registrada:
                              type: string
                              format: date-time
                              pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d:[0-5]\d(?:\.\d+)?(?:Z|([+-](?:[01]\d|2[0-3]):[0-5]\d)))$
                            nombre:
                              type:
                                - string
                                - "null"
                            correo:
                              type:
                                - string
                                - "null"
                            correoVerificado:
                              anyOf:
                                - type: string
                                  format: date-time
                                  pattern: ^(?:(?:\d\d[2468][048]|\d\d[13579][26]|\d\d0[48]|[02468][048]00|[13579][26]00)-02-29|\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\d|30)|(?:02)-(?:0[1-9]|1\d|2[0-8])))T(?:(?:[01]\d|2[0-3]):[0-5]\d:[0-5]\d(?:\.\d+)?(?:Z|([+-](?:[01]\d|2[0-3]):[0-5]\d)))$
                                - type: "null"
                          required:
                            - registrada
                            - nombre
                            - correo
                            - correoVerificado
                          additionalProperties: false
                      required:
                        - id
                        - petId
                        - reclamantePetId
                        - clinicId
                        - estado
                        - motivo
                        - plazo
                        - creada
                        - resuelta
                        - nota
                        - clinica
                        - titular
                        - reclamante
                      additionalProperties: false
                required:
                  - reclamaciones
                additionalProperties: false
        "404":
          description: Error 404
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      x-acceso: operador
  /ops/v1/claims/{id}/resolve:
    post:
      tags:
        - Operador
      summary: Resolver a mano, con una nota que queda guardada y avisa a las dos partes. Una abierta en plazo no se adelanta.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      security: *a5
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $schema: https://json-schema.org/draft/2020-12/schema
              type: object
              properties:
                aFavor:
                  type: string
                  enum:
                    - reclamante
                    - titular
                nota:
                  type: string
                  minLength: 5
                  maxLength: 2000
              required:
                - aFavor
                - nota
      responses:
        "200":
          description: Correcto
          content:
            application/json:
              schema:
                $schema: https://json-schema.org/draft/2020-12/schema
                type: object
                properties:
                  ok:
                    type: boolean
                    const: true
                  estado:
                    type: string
                    enum:
                      - a-favor-reclamante
                      - a-favor-titular
                required:
                  - ok
                  - estado
                additionalProperties: false
        "400":
          description: Error 400
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Error 404
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "409":
          description: Error 409
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
      x-acceso: operador
components:
  securitySchemes:
    sesionDueno:
      type: apiKey
      in: cookie
      name: bam_owner
      description: Portal del dueño.
    tokenApp:
      type: http
      scheme: bearer
      description: "App: token de sesión, con la cabecera `x-bm-cliente: app`."
    sesionClinica:
      type: apiKey
      in: cookie
      name: bam_clinic
      description: Identifica a la organización y a la persona. No es una llave a los datos.
    claveApi:
      type: http
      scheme: bearer
      description: "Software de gestión: `bmk_…`."
    tokenOperador:
      type: http
      scheme: bearer
      description: "Panel del operador: OPS_TOKEN."
  schemas:
    Error:
      $schema: https://json-schema.org/draft/2020-12/schema
      type: object
      properties:
        error:
          type: string
        motivo:
          type: string
      required:
        - error
      additionalProperties: {}